+++ src/test/java/blackjack/card/CardTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
